feat: loads all unnecessary buttons lazily
